requirements.txt
Flask==3.0.3
google-generativeai==0.3.2
python-dotenv==1.0.1
fpdf==1.7.2
gunicorn==23.0.0
